test-mongoc-mongohouse.c: address C4244 narrowing conversion warnings

+++ src/libmongoc/tests/test-mongoc-mongohouse.c
@@ -41,7 +41,7 @@ test_mongohouse_cb (bson_t *scenario)
 typedef struct {
    /* Information from original cursor */
    char *cursor_ns;
-   int cursor_id;
+   int64_t cursor_id;
 
    bool parsed_cursor;
    bool parsed_cmd_started;
@@ -52,7 +52,7 @@ typedef struct {
 
 
 static bool
-cursor_in_killed_array (bson_t *cursors_killed, int cursor_id)
+cursor_in_killed_array (bson_t *cursors_killed, int64_t cursor_id)
 {
    bson_iter_t iter;
